desafio-ada-back/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.6
django-filter==22.1
django-import-export==3.0.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.2.0
ipdb==0.13.11
ipython==8.9.0
jedi==0.18.2
MarkupPy==1.14
matplotlib-inline==0.1.6
odfpy==1.4.1
openpyxl==3.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tablib==3.3.0
traitlets==5.9.0
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
